+++ src/plplot/owl_plplot.ml
@@ -1,1 +1,0 @@
-module Plot = Owl_plot