+++ test/test_runner.ml
@@ -9,8 +9,6 @@ let () =
     ; "algodiff grad", Unit_algodiff_grad.test_set
     ; "dense matrix", Unit_dense_matrix.test_set
     ; "dense ndarray", Unit_dense_ndarray.test_set
-    ; "sparse matrix", Unit_sparse_matrix.test_set
-    ; "sparse ndarray", Unit_sparse_ndarray.test_set
     ; "ndarray core", Unit_ndarray_core.test_set
     ; "ndarray primitive", Unit_ndarray_primitive.test_set
     ; "ndarray operation", Unit_ndarray_operation.test_set